+++ examples/query-light-with-react-router-v6/src/App.tsx
@@ -1,6 +1,6 @@
 import { Todo as TodoType } from "./types";
 import Todo from "./components/Todo";
-import { useQueryLight } from "@rime-ui/react-query-light";
+import { useQueryLight } from "@rime-ui/query-light";
 function App() {
   const { data: todos, isLoading } = useQueryLight<TodoType[]>(
     ["todos"],