+++ packages/query-light-core/src/index.ts
@@ -1,4 +1,4 @@
-export type * from "./types.js";
-export * from "./queryCache.js";
-export * from "./retryer.js";
-export * from "./graphql.js";
+export type * from "./types";
+export * from "./queryCache";
+export * from "./retryer";
+export * from "./graphql";