+++ packages/query-light/src/index.ts
@@ -1,2 +1,2 @@
-export * from "@rime-ui/query-light-core"
+export * from "@rime-ui/query-light-main-core"
 export { QueryCache, type QueryCacheType, createGraphQLMutation, createGraphQLQuery, createGraphQLTypedMutation, createGraphQLTypedQuery, QueryLightProvider, useMutationLight, useQueryCache, useQueryLight } from "@rime-ui/react-query-light"