+++ packages/react-query-light/src/index.ts
@@ -2,12 +2,12 @@ export { useQueryLight } from "./useQueryLight";
 export { useMutationLight } from "./useMutationLight";
 export { QueryLightProvider, useQueryCache } from "./QueryLightProvider";
 
-export { QueryCache } from "@rime-ui/query-light-core";
-export type { QueryCacheType } from "@rime-ui/query-light-core";
+export { QueryCache } from "@rime-ui/query-light-main-core";
+export type { QueryCacheType } from "@rime-ui/query-light-main-core";
 
 export {
   createGraphQLQuery,
   createGraphQLMutation,
   createGraphQLTypedQuery,
   createGraphQLTypedMutation,
-} from "@rime-ui/query-light-core";
+} from "@rime-ui/query-light-main-core";