+++ packages/query-light/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rime-ui/query-light
 
+## 2.0.2
+
+### Patch Changes
+
+- fix version issue
+- Updated dependencies
+  - @rime-ui/react-query-light@1.0.3
+  - @rime-ui/query-light-main-core@1.0.1
+
 ## 2.0.0
 
 ### Major Changes